refactor(utils/misc): move get-data type functions into separate file & add docstrings     for #5187
