+++ frontend/src/ts/elements/keymap.ts
@@ -4,6 +4,7 @@ import * as SlowTimer from "../states/slow-timer";
 import * as ConfigEvent from "../observables/config-event";
 import * as KeymapEvent from "../observables/keymap-event";
 import * as Misc from "../utils/misc";
+import * as GetData from "../utils/get-data";
 import * as Hangul from "hangul-js";
 import * as Notifications from "../elements/notifications";
 import * as ActivePage from "../states/active-page";
@@ -115,7 +116,7 @@ export async function refresh(
   try {
     let layouts;
     try {
-      layouts = await Misc.getLayoutsList();
+      layouts = await GetData.getLayoutsList();
     } catch (e) {
       Notifications.add(
         Misc.createErrorMessage(e, "Failed to refresh keymap"),
