+++ frontend/src/ts/pages/settings.ts
@@ -2,6 +2,7 @@ import SettingsGroup from "../settings/settings-group";
 import Config, * as UpdateConfig from "../config";
 import * as Sound from "../controllers/sound-controller";
 import * as Misc from "../utils/misc";
+import * as GetData from "../utils/get-data";
 import * as DB from "../db";
 import { toggleFunbox } from "../test/funbox/funbox";
 import * as TagController from "../controllers/tag-controller";
@@ -442,7 +443,7 @@ async function fillSettingsPage(): Promise<void> {
 
   let languageGroups;
   try {
-    languageGroups = await Misc.getLanguageGroups();
+    languageGroups = await GetData.getLanguageGroups();
   } catch (e) {
     console.error(
       Misc.createErrorMessage(
@@ -478,7 +479,7 @@ async function fillSettingsPage(): Promise<void> {
 
   let layoutsList;
   try {
-    layoutsList = await Misc.getLayoutsList();
+    layoutsList = await GetData.getLayoutsList();
   } catch (e) {
     console.error(Misc.createErrorMessage(e, "Failed to refresh keymap"));
   }
@@ -521,7 +522,7 @@ async function fillSettingsPage(): Promise<void> {
 
   let themes;
   try {
-    themes = await Misc.getThemesList();
+    themes = await GetData.getThemesList();
   } catch (e) {
     console.error(
       Misc.createErrorMessage(e, "Failed to load themes into dropdown boxes")
@@ -581,7 +582,7 @@ async function fillSettingsPage(): Promise<void> {
 
   let funboxList;
   try {
-    funboxList = await Misc.getFunboxList();
+    funboxList = await GetData.getFunboxList();
   } catch (e) {
     console.error(Misc.createErrorMessage(e, "Failed to get funbox list"));
   }
@@ -630,7 +631,7 @@ async function fillSettingsPage(): Promise<void> {
 
   let fontsList;
   try {
-    fontsList = await Misc.getFontsList();
+    fontsList = await GetData.getFontsList();
   } catch (e) {
     console.error(
       Misc.createErrorMessage(e, "Failed to update fonts settings buttons")
@@ -787,7 +788,7 @@ function setActiveFunboxButton(): void {
   $(`.pageSettings .section[data-config-name='funbox'] .button`).removeClass(
     "disabled"
   );
-  Misc.getFunboxList()
+  GetData.getFunboxList()
     .then((funboxModes) => {
       funboxModes.forEach((funbox) => {
         if (
