+++ frontend/src/ts/utils/misc.ts
@@ -1,276 +1,7 @@
 import * as Loader from "../elements/loader";
 import { envConfig } from "../constants/env-config";
-import { hexToHSL } from "./colors";
 import { roundTo2, randomIntFromRange } from "./numbers";
-
-async function fetchJson<T>(url: string): Promise<T> {
-  try {
-    if (!url) throw new Error("No URL");
-    const res = await fetch(url);
-    if (res.ok) {
-      return await res.json();
-    } else {
-      throw new Error(`${res.status} ${res.statusText}`);
-    }
-  } catch (e) {
-    console.error("Error fetching JSON: " + url, e);
-    throw e;
-  }
-}
-
-export const cachedFetchJson = memoizeAsync<string, typeof fetchJson>(
-  fetchJson
-);
-
-export async function getLayoutsList(): Promise<MonkeyTypes.Layouts> {
-  try {
-    const layoutsList = await cachedFetchJson<MonkeyTypes.Layouts>(
-      "/layouts/_list.json"
-    );
-    return layoutsList;
-  } catch (e) {
-    throw new Error("Layouts JSON fetch failed");
-  }
-}
-
-/**
- * @throws {Error} If layout list or layout doesnt exist.
- */
-export async function getLayout(
-  layoutName: string
-): Promise<MonkeyTypes.Layout> {
-  const layouts = await getLayoutsList();
-  const layout = layouts[layoutName];
-  if (layout === undefined) {
-    throw new Error(`Layout ${layoutName} is undefined`);
-  }
-  return layout;
-}
-
-let themesList: MonkeyTypes.Theme[] | undefined;
-export async function getThemesList(): Promise<MonkeyTypes.Theme[]> {
-  if (!themesList) {
-    let themes = await cachedFetchJson<MonkeyTypes.Theme[]>(
-      "/themes/_list.json"
-    );
-
-    themes = themes.sort(function (a: MonkeyTypes.Theme, b: MonkeyTypes.Theme) {
-      const nameA = a.name.toLowerCase();
-      const nameB = b.name.toLowerCase();
-      if (nameA < nameB) return -1;
-      if (nameA > nameB) return 1;
-      return 0;
-    });
-    themesList = themes;
-    return themesList;
-  } else {
-    return themesList;
-  }
-}
-
-let sortedThemesList: MonkeyTypes.Theme[] | undefined;
-export async function getSortedThemesList(): Promise<MonkeyTypes.Theme[]> {
-  if (!sortedThemesList) {
-    if (!themesList) {
-      await getThemesList();
-    }
-    if (!themesList) {
-      throw new Error("Themes list is undefined");
-    }
-    let sorted = [...themesList];
-    sorted = sorted.sort((a, b) => {
-      const b1 = hexToHSL(a.bgColor);
-      const b2 = hexToHSL(b.bgColor);
-      return b2.lgt - b1.lgt;
-    });
-    sortedThemesList = sorted;
-    return sortedThemesList;
-  } else {
-    return sortedThemesList;
-  }
-}
-
-export async function getLanguageList(): Promise<string[]> {
-  try {
-    const languageList = await cachedFetchJson<string[]>(
-      "/languages/_list.json"
-    );
-    return languageList;
-  } catch (e) {
-    throw new Error("Language list JSON fetch failed");
-  }
-}
-
-export async function getLanguageGroups(): Promise<
-  MonkeyTypes.LanguageGroup[]
-> {
-  try {
-    const languageGroupList = await cachedFetchJson<
-      MonkeyTypes.LanguageGroup[]
-    >("/languages/_groups.json");
-    return languageGroupList;
-  } catch (e) {
-    throw new Error("Language groups JSON fetch failed");
-  }
-}
-
-let currentLanguage: MonkeyTypes.LanguageObject;
-export async function getLanguage(
-  lang: string
-): Promise<MonkeyTypes.LanguageObject> {
-  // try {
-  if (currentLanguage === undefined || currentLanguage.name !== lang) {
-    currentLanguage = await cachedFetchJson<MonkeyTypes.LanguageObject>(
-      `/languages/${lang}.json`
-    );
-  }
-  return currentLanguage;
-  // } catch (e) {
-  //   console.error(`error getting language`);
-  //   console.error(e);
-  //   currentLanguage = await cachedFetchJson<MonkeyTypes.LanguageObject>(
-  //     `/language/english.json`
-  //   );
-  //   return currentLanguage;
-  // }
-}
-
-export async function getCurrentLanguage(
-  languageName: string
-): Promise<MonkeyTypes.LanguageObject> {
-  return await getLanguage(languageName);
-}
-
-export async function getCurrentGroup(
-  language: string
-): Promise<MonkeyTypes.LanguageGroup | undefined> {
-  let retgroup: MonkeyTypes.LanguageGroup | undefined;
-  const groups = await getLanguageGroups();
-  groups.forEach((group) => {
-    if (retgroup === undefined) {
-      if (group.languages.includes(language)) {
-        retgroup = group;
-      }
-    }
-  });
-  return retgroup;
-}
-
-let funboxList: MonkeyTypes.FunboxMetadata[] | undefined;
-export async function getFunboxList(): Promise<MonkeyTypes.FunboxMetadata[]> {
-  if (!funboxList) {
-    let list = await cachedFetchJson<MonkeyTypes.FunboxMetadata[]>(
-      "/funbox/_list.json"
-    );
-    list = list.sort(function (
-      a: MonkeyTypes.FunboxMetadata,
-      b: MonkeyTypes.FunboxMetadata
-    ) {
-      const nameA = a.name.toLowerCase();
-      const nameB = b.name.toLowerCase();
-      if (nameA < nameB) return -1;
-      if (nameA > nameB) return 1;
-      return 0;
-    });
-    funboxList = list;
-    return funboxList;
-  } else {
-    return funboxList;
-  }
-}
-
-export async function getFunbox(
-  funbox: string
-): Promise<MonkeyTypes.FunboxMetadata | undefined> {
-  const list: MonkeyTypes.FunboxMetadata[] = await getFunboxList();
-  return list.find(function (element) {
-    return element.name === funbox;
-  });
-}
-
-let fontsList: MonkeyTypes.FontObject[] | undefined;
-export async function getFontsList(): Promise<MonkeyTypes.FontObject[]> {
-  if (!fontsList) {
-    let list = await cachedFetchJson<MonkeyTypes.FontObject[]>(
-      "/fonts/_list.json"
-    );
-    list = list.sort(function (
-      a: MonkeyTypes.FontObject,
-      b: MonkeyTypes.FontObject
-    ) {
-      const nameA = a.name.toLowerCase();
-      const nameB = b.name.toLowerCase();
-      if (nameA < nameB) return -1;
-      if (nameA > nameB) return 1;
-      return 0;
-    });
-    fontsList = list;
-    return fontsList;
-  } else {
-    return fontsList;
-  }
-}
-
-export async function getChallengeList(): Promise<MonkeyTypes.Challenge[]> {
-  try {
-    const data = await cachedFetchJson<MonkeyTypes.Challenge[]>(
-      "/challenges/_list.json"
-    );
-    return data;
-  } catch (e) {
-    throw new Error("Challenge list JSON fetch failed");
-  }
-}
-
-export async function getSupportersList(): Promise<string[]> {
-  try {
-    const data = await cachedFetchJson<string[]>("/about/supporters.json");
-    return data;
-  } catch (e) {
-    throw new Error("Supporters list JSON fetch failed");
-  }
-}
-
-export async function getContributorsList(): Promise<string[]> {
-  try {
-    const data = await cachedFetchJson<string[]>("/about/contributors.json");
-    return data;
-  } catch (e) {
-    throw new Error("Contributors list JSON fetch failed");
-  }
-}
-
-export async function getLatestReleaseFromGitHub(): Promise<string> {
-  type releaseType = { name: string };
-  const releases = await cachedFetchJson<releaseType[]>(
-    "https://api.github.com/repos/monkeytypegame/monkeytype/releases?per_page=1"
-  );
-  if (releases[0] === undefined || releases[0].name === undefined) {
-    throw new Error("No release found");
-  }
-  return releases[0].name;
-}
-
-export async function getReleasesFromGitHub(): Promise<
-  MonkeyTypes.GithubRelease[]
-> {
-  return cachedFetchJson(
-    "https://api.github.com/repos/monkeytypegame/monkeytype/releases?per_page=5"
-  );
-}
-
-// function getPatreonNames() {
-//   let namesel = $(".pageAbout .section .supporters");
-//   firebase
-//     .functions()
-//     .httpsCallable("getPatreons")()
-//     .then((data) => {
-//       let names = data.data;
-//       names.forEach((name) => {
-//         namesel.append(`<div>${name}</div>`);
-//       });
-//     });
-// }
+import * as GetData from "./get-data";
 
 export function getLastChar(word: string): string {
   try {
@@ -1035,14 +766,6 @@ export function isAnyPopupVisible(): boolean {
   return popupVisible;
 }
 
-export function getLevel(xp: number): number {
-  return (1 / 98) * (-151 + Math.sqrt(392 * xp + 22801)) + 1;
-}
-
-export function getXpForLevel(level: number): number {
-  return 49 * (level - 1) + 100;
-}
-
 export async function promiseAnimation(
   el: JQuery,
   animation: Record<string, string>,
@@ -1168,7 +891,7 @@ export function dreymarIndex(arrayLength: number): number {
 export async function checkIfLanguageSupportsZipf(
   language: string
 ): Promise<"yes" | "no" | "unknown"> {
-  const lang = await getLanguage(language);
+  const lang = await GetData.getLanguage(language);
   if (lang.orderedByFrequency === true) return "yes";
   if (lang.orderedByFrequency === false) return "no";
   return "unknown";
